requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
astor==0.8.1
Babel==2.15.0
bce-python-sdk==0.9.23
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
crashtest==0.4.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
easydict==1.13
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.7.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.54.1
frontend==0.0.3
future==1.0.0
gast==0.3.3
ghp-import==2.1.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
installer==0.7.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
keyring==24.3.1
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
looseversion==1.3.0
lxml==5.3.0
Markdown==3.6
MarkupSafe==3.0.1
matplotlib==3.9.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.29
mkdocs-material-extensions==1.3.1
mkdocs-table-reader-plugin==2.2.2
more-itertools==10.3.0
msgpack==1.0.8
networkx==3.2.1
nibabel==5.3.0
nipype==1.8.6
numpy==1.26.4
openai==1.51.2
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
paddleclas==2.5.2
paddleocr==2.8.1
paddlepaddle==2.6.2
paddlepaddle-gpu==2.6.2
paginate==0.5.6
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
premailer==3.10.0
prettytable==3.11.0
protobuf==5.28.2
prov==2.0.1
psutil==6.0.0
ptyprocess==0.7.0
pyclipper==1.3.0.post5
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
pymdown-extensions==10.8.1
PyMuPDF==1.24.11
pyparsing==3.1.4
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
rapidfuzz==3.9.4
rarfile==4.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
SecretStorage==3.3.3
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.39.2
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.8.0
tomli==2.0.1
tqdm==4.66.5
traits==6.3.2
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.26.3
visualdl==2.5.3
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
